requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
editdistpy==0.1.6
email_validator==2.2.0
fastapi==0.120.2
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Levenshtein==0.27.3
numpy<2.0.0,>=1.24.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pillow_heif==1.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
easyocr==1.7.1
torch>=2.0.0
torchvision>=0.15.0
python-dotenv==1.2.1
python-jose==3.5.0
python-Levenshtein==0.27.3
python-multipart==0.0.20
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.1
symspellpy==6.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions>=4.9.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
